Rename gps to location & gps task to capture location task
