+++ ground/src/main/res/layout/capture_location_task_frag.xml
@@ -21,7 +21,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.datacollection.tasks.gps.GpsTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.location.CaptureLocationTaskViewModel" />
   </data>
 
   <TextView
